fix: brought back App.tsx, separated theme
Related work item: GH-6
