+++ tsconfig.json
@@ -30,5 +30,5 @@
     "forceConsistentCasingInFileNames": true,
     "jsx": "react-jsx"
   },
-  "include": ["src/**/*"]
+  "include": ["./src"]
 }
